backend/pyproject.toml
[project]
name = "recipe-log"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.3.0",
    "pydantic[email]>=2.0.0",
    "recipe-scrapers>=14.0.0",
    "beautifulsoup4>=4.12.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "isodate>=0.6.1",
    "pillow-heif>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
